Am getting tired of those who can't follow the Standard.
Caleb, use Constants and Constant files ALWAYS when we are on a Code_2022 repo.  I dislike magic numbers with a passion akin to when Khalessi burnt Kings Landing.  Stop doing it /your/ way, and follow the Standard, Caleb.  The standard exists for a reason.  The best coders /follow/ it for a reason.  Next time I discover you not following the Standard, YOU sir will be the one who has to integrate merges.